4,26

+++ LicentaApp/Models/Category.cs
@@ -13,12 +13,12 @@ namespace LicentaApp.Models
         }
 
         //adaugam constructorul clasei
-        public Category(short id, string nume)
+        public Category(int id, string nume)
         {
             Id = id;
             Nume = nume;
         }
-        public short Id { get; set; }
+        public int Id { get; set; }
         public string  Nume { get; set; }
     }
 }
